Refactor file attachment handling into new component
